refactor: rename delete-permanently-item.tsx to use-delete-permanently-item.tsx
