+++ src/views/contact-groups/actions/delete-contact-group.tsx
@@ -9,8 +9,8 @@ import { useSnackbar } from '@zextras/carbonio-design-system';
 import { closeBoard, getBoardById } from '@zextras/carbonio-shell-ui';
 import { useTranslation } from 'react-i18next';
 
-import { useDeletePermanentlyItem } from '../../../actions/delete-permanently-item';
 import { UIAction } from '../../../actions/types';
+import { useDeletePermanentlyItem } from '../../../actions/use-delete-permanently-item';
 import { EDIT_CONTACT_GROUP_BOARD_ID } from '../../../constants';
 import { useGetContactGroupFromPath } from '../../../hooks/useGetContactGroupFromPath';
 import { ContactGroup } from '../../../model/contact-group';
